feat: refactor filter text parsing logic and badge creation in NetworkFilterProvider
